Expose mnode_print_stat as a public function and remove obsolete debug macros.
- Refactor mnode_print_stat to accept `mtree *mt` directly with NULL
  safety checks.
- Remove `MNODE_DEBUG_STAT` macro guard to make statistics collection
  always available.
- Extract `struct migemo` definition into internal `migemo_struct.h` header.
- Update profile_speed test to print mnode statistics upon loading the
  dictionary.
